feat: hurry cache show cargo
