+++ packages/hurry/src/cargo.rs
@@ -28,9 +28,7 @@ mod workspace;
 pub use build_args::{CargoBuildArgument, CargoBuildArguments, ColorWhen, MessageFormat};
 pub use build_plan::BuildPlan;
 pub use build_script::{BuildScriptOutput, RootOutput};
-pub use cache::{
-    ArtifactKey, ArtifactPlan, BuiltArtifact, CacheStats, CargoCache, LibraryUnitHash, RestoreState,
-};
+pub use cache::{CacheStats, CargoCache, RestoreState};
 pub use dep_info::{DepInfo, DepInfoLine};
 pub use dependency::{Dependency, DependencyBuild, Optimizations};
 pub use path::QualifiedPath;
@@ -43,7 +41,9 @@ pub use unit_graph::{
     CargoCompileMode, UnitGraph, UnitGraphDependency, UnitGraphProfile,
     UnitGraphProfilePanicStrategy, UnitGraphUnit,
 };
-pub use workspace::*;
+pub use workspace::{
+    ArtifactKey, ArtifactPlan, BuildScriptDirs, BuiltArtifact, LibraryUnitHash, Workspace,
+};
 
 /// Execute Cargo without a subcommand with specified arguments.
 #[instrument]
